refactor rescue and storage cli to include as subcommand in libra-cli
