+++ tools/rescue/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod diem_db_bootstrapper;
 pub mod rescue_tx;
+pub mod rescue_cli;
 pub mod session_tools;
 pub mod twin;
 pub use twin::{Twin, TwinSetup};
