+++ tools/rescue/tests/rescue_cli_creates_blob.rs
@@ -1,7 +1,7 @@
 mod support;
 
 use libra_smoke_tests::libra_smoke::LibraSmoke;
-use rescue::{diem_db_bootstrapper::BootstrapOpts, rescue_tx::RescueTxOpts};
+use libra_rescue::{diem_db_bootstrapper::BootstrapOpts, rescue_tx::RescueTxOpts};
 
 #[tokio::test]
 async fn test_valid_genesis() -> anyhow::Result<()> {
