+++ tools/rescue/tests/upgrade_multi_action.rs
@@ -7,7 +7,7 @@ use libra_txs::{
     txs_cli_community::{CageTx, ClaimTx, CommunityTxs, InitTx},
 };
 use libra_types::core_types::app_cfg::TxCost;
-use rescue::twin::{Twin, TwinSetup};
+use libra_rescue::twin::{Twin, TwinSetup};
 use std::{env, path::PathBuf};
 
 // TODO: Remove after offer structure is migrated
