+++ tools/storage/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod read_snapshot;
+pub mod storage_cli;
